+++ src/api/entities/types.ts
@@ -12,6 +12,7 @@ import {
   Portfolio as PortfolioClass,
   // Proposal as ProposalClass,
   SecurityToken as SecurityTokenClass,
+  Sto as StoClass,
   TickerReservation as TickerReservationClass,
   Venue as VenueClass,
 } from '~/internal';
@@ -28,6 +29,7 @@ export type Instruction = InstanceType<typeof InstructionClass>;
 export type Portfolio = InstanceType<typeof PortfolioClass>;
 export type DefaultPortfolio = InstanceType<typeof DefaultPortfolioClass>;
 export type NumberedPortfolio = InstanceType<typeof NumberedPortfolioClass>;
+export type Sto = InstanceType<typeof StoClass>;
 // export type Proposal = InstanceType<typeof ProposalClass>;
 
 export * from './TickerReservation/types';
@@ -35,4 +37,5 @@ export * from './SecurityToken/types';
 export * from './Venue/types';
 export * from './Instruction/types';
 export * from './Portfolio/types';
+export * from './Sto/types';
 // export * from './Proposal/types';
